docsmart_django/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chargebee==2.8.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.1.7
django-appconf==1.0.4
django-cors-headers==3.7.0
django-cryptography==1.0
django-filter==2.4.0
django-heroku==0.3.1
django-mdeditor==0.1.18
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.2.0
drf-yasg==1.20.0
idna==2.10
importlib-metadata==3.8.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
nested-lookup==0.2.22
openapi-codec==1.3.2
packaging==20.9
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.2
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
zipp==3.4.1